requirements.txt
tensorflow
flask
numpy
Pillow
flask-sqlalchemy